Changed sensing of bool attr overwrite
We used this so far:

```ts
if (attributes.isLive === true || attributes.isLive === false) {
```

This changed by ESLint's `--fix` to:

```ts
if (attributes.isLive) {
```

Not what we intended. This following should ensure that we are
setting a boolean property only when it was submitted by
the `attributes` arguments:

```ts
if (attributes.isLive !== undefined) {
```

This still changes the behavior of a property that is set to `null`.